+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- (PlusIcon): Fix SVG attribute casing
 - (InputField): Fix props not being passed to inner input element
 
 # v2.0.0-dev.2